Move openai client to fixture
